refactor(radiant): split light-dom shim into minimal-dom modules
